Use typescript in content scripts
