+++ src/content-scripts/modules/icons/index.ts
