+++ src/content-scripts/modules/utilities/addStyles.ts
