+++ src/content-scripts/modules/utilities/constructNewData.ts
