+++ src/content-scripts/modules/utilities/math.ts
