+++ src/content-scripts/modules/utilities/pageUtils.ts
