+++ src/content-scripts/modules/utilities/removeElement.ts
