+++ src/content-scripts/modules/utilities/storage.ts
