--- src/content-scripts/selectors.js
@@ -1,15 +1,0 @@
-const selectors = {};
-
-// Layout
-
-/* Some examples
-
-selectors.mainWrapper = `main[role="main"]`;
-selectors.sidebarLinks = {
-  home: `[data-testid="AppTabBar_Home_Link"]`,
-  explore: `[data-testid="AppTabBar_Explore_Link"]`,
-};
-
-*/
-
-export default selectors;
